chore(support-bot): rewrite with opencode and remote GitHub MCP
