+++ sample-code/src/support-bot/package.json
@@ -1,0 +1,24 @@
+{
+  "type": "module",
+  "name": "support-bot",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "local": "node --env-file=.env reply.ts"
+  },
+  "dependencies": {
+    "@opencode-ai/sdk": "^1.18.4",
+    "opencode-ai": "^1.18.4"
+  },
+  "devDependencies": {
+    "@types/node": "^22.0.0",
+    "typescript": "^5.0.0"
+  },
+  "engines": {
+    "node": ">=22"
+  },
+  "packageManager": "pnpm@10.1.0",
+  "pnpm": {
+    "onlyBuiltDependencies": ["opencode-ai"]
+  }
+}
